OJ-2813: Refactor canonical Address type
Move out canonical address type into it's own
file i.e canonical-address.ts

Restructure type in address.ts ready for addresses-v2

Update existing addresses code to make use of canonical-address
type
