+++ lambdas/get-addresses/src/types/address.ts
@@ -1,17 +1,6 @@
-export type CanonicalAddress = {
-    validFrom?: string;
-    validUntil?: string;
-    uprn?: string;
-    organisationName?: string;
-    departmentName?: string;
-    subBuildingName?: string;
-    buildingName?: string;
-    buildingNumber?: string;
-    dependentStreetName?: string;
-    streetName?: string;
-    doubleDependentAddressLocality?: string;
-    dependentAddressLocality?: string;
-    addressLocality?: string;
-    postalCode?: string;
-    addressCountry?: string;
+import { CanonicalAddress } from "./canonical-address";
+
+export type Address = {
+    addresses: CanonicalAddress[];
+    context?: string;
 };
