+++ lambdas/get-addresses/src/types/canonical-address.ts
@@ -1,0 +1,17 @@
+export type CanonicalAddress = {
+    validFrom?: string;
+    validUntil?: string;
+    uprn?: string;
+    organisationName?: string;
+    departmentName?: string;
+    subBuildingName?: string;
+    buildingName?: string;
+    buildingNumber?: string;
+    dependentStreetName?: string;
+    streetName?: string;
+    doubleDependentAddressLocality?: string;
+    dependentAddressLocality?: string;
+    addressLocality?: string;
+    postalCode?: string;
+    addressCountry?: string;
+};
